# tools/sparkbox/internal/restapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "sparkbox control API",
    "version": "1.0.0",
    "summary": "Everything `ssh ctl@<domain>` does, over HTTPS.",
...
    "contact": {
      "name": "sparkbox",
      "url": "https://example.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "This host."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "operationId": "capabilities",
        "tags": ["account"],
        "summary": "What this host has configured",
        "description": "Archiving needs object storage, snapshots need a driver that supports them, scheduling and tagging need their stores. Read this instead of discovering a 501 by trying.",
        "responses": {
          "200": {
            "description": "The host's feature set.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Capabilities" },
                "example": {
                  "archiving": true,
                  "snapshots": true,
                  "scheduling": true,
                  "tags": true,
                  "routes": true,
                  "session_tokens": true,
                  "terminal": true
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/whoami": {
      "get": {
        "operationId": "whoami",
        "tags": ["account"],
        "summary": "The authenticated account",
        "description": "Mirrors `ssh ctl@example.com whoami`. `key_fp` is empty here: an HTTP request carries no SSH key.",
        "responses": {
          "200": {
            "description": "Your account.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Whoami" },
                "example": {
                  "handle": "vanpelt",
                  "status": "active",
                  "operator": true,
                  "email": "me@example.com",
                  "github_login": "vanpelt",
                  "subject": "e3b0c44298fc1c14"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/sandboxes": {
      "get": {
        "operationId": "list",
        "tags": ["sandboxes"],
        "summary": "List your sandboxes",
        "description": "Every sandbox you own, in creation order. Other owners' sandboxes are not listed and not counted.",
        "responses": {
          "200": {
            "description": "Your sandboxes.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SandboxList" },
                "example": {
                  "sandboxes": [
                    {
                      "name": "dazzling-canyon",
                      "owner": "vanpelt",
                      "state": "running",
                      "pinned": false,
                      "tags": ["prod"],
                      "vcpus": 2,
                      "mem_mb": 8192,
                      "disk_mb": 25600,
                      "created_at": "2026-07-19T10:04:11Z",
                      "last_active": "2026-07-21T18:22:03Z",
                      "url": "https://dazzling-canyon.example.com",
                      "terminal_url": "https://dazzling-canyon-xterm.example.com"
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "create",
        "tags": ["sandboxes"],
        "summary": "Create a sandbox",
        "description": "The REST form of `ssh new@example.com`. Omit `name` and the platform generates an adjective-noun one.\n\nTags are stamped **before** the VM is created, because creation fires the secret-environment push asynchronously and the tags decide its contents. If creation fails, the tag rows are rolled back.\n\nUnlike the SSH door, there is no way to pass a command: named fields have no ambiguity to resolve.",
        "parameters": [
          { "$ref": "#/components/parameters/Prefer" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateRequest" },
              "example": { "name": "demo", "tags": ["prod", "eu"] }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The sandbox is up.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sandbox" }
              }
            }
          },
          "202": { "$ref": "#/components/responses/Accepted" },
          "400": { "$ref": "#/components/responses/Invalid" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "No machine in this fleet is called `node`. Node names are not masked: an operator publishes them so callers can place work.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "409": {
            "description": "That name is taken or reserved.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RunningLimit" },
          "503": { "$ref": "#/components/responses/HostAtCapacity" },
          "507": { "$ref": "#/components/responses/DiskPoolFull" }
        }
      }
    },
    "/v1/sandboxes/{name}": {
      "get": {
        "operationId": "get",
        "tags": ["sandboxes"],
        "summary": "Read one sandbox",
        "description": "Does not resume a paused sandbox — reading is never a side effect.",
        "parameters": [{ "$ref": "#/components/parameters/SandboxName" }],
        "responses": {
          "200": {
            "description": "The sandbox.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sandbox" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "rm",
        "tags": ["sandboxes"],
        "summary": "Destroy a sandbox",
        "description": "Irreversible. Cascades to the sandbox's web routes, schedules, tags, front-door DNS record and, if it was archived, the stored object.",
        "parameters": [
          { "$ref": "#/components/parameters/SandboxName" },
          { "$ref": "#/components/parameters/Prefer" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": {
            "description": "Destroyed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Deleted" },
                "example": { "name": "demo", "deleted": true }
              }
            }
          },
          "202": { "$ref": "#/components/responses/Accepted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/sandboxes/{name}/pause": {
      "post": {
        "operationId": "pause",
        "tags": ["sandboxes"],
        "summary": "Pause a sandbox",
        "description": "Writes the guest's full memory snapshot and frees its RAM. Resuming restores the running processes exactly. Budget: 3 minutes.",
        "parameters": [
          { "$ref": "#/components/parameters/SandboxName" },
          { "$ref": "#/components/parameters/Prefer" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": {
            "description": "Paused.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sandbox" }
              }
            }
          },
          "202": { "$ref": "#/components/responses/Accepted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/sandboxes/{name}/resume": {
      "post": {
        "operationId": "restore",
        "tags": ["sandboxes"],
        "summary": "Resume (or restore) a sandbox",
        "description": "Starts a paused sandbox from its memory snapshot. If the sandbox was archived, this also downloads and decompresses its disk first, which is why the budget is 15 minutes.",
        "parameters": [
          { "$ref": "#/components/parameters/SandboxName" },
          { "$ref": "#/components/parameters/Prefer" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": {
            "description": "Running.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sandbox" }
              }
            }
          },
          "202": { "$ref": "#/components/responses/Accepted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RunningLimit" },
          "503": { "$ref": "#/components/responses/HostAtCapacity" }
        }
      }
    },
    "/v1/sandboxes/{name}/archive": {
      "post": {
        "operationId": "archive",
        "tags": ["sandboxes"],
        "summary": "Archive a sandbox to object storage",
        "description": "Checks the filesystem, zeroes free space, compresses the disk and uploads it, then frees the host disk entirely. Resuming restores it. Needs object storage: hosts without it answer 501. Budget: 15 minutes, so this almost always returns 202.",
        "parameters": [
          { "$ref": "#/components/parameters/SandboxName" },
          { "$ref": "#/components/parameters/Prefer" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": {
            "description": "Archived.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sandbox" }
              }
            }
          },
          "202": { "$ref": "#/components/responses/Accepted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "501": { "$ref": "#/components/responses/Disabled" }
        }
      }
    },
    "/v1/sandboxes/{name}/resize": {
      "post": {
        "operationId": "resize",
        "tags": ["sandboxes"],
        "summary": "Grow the root disk",
        "description": "Grow-only. The sandbox is paused, its memory snapshot is **discarded**, the filesystem is checked and grown, and the guest cold-boots — so anything running inside dies. Send `size` (\"25G\", \"512M\", or a bare number meaning GB) or `size_mb`. Budget: 10 minutes.",
        "parameters": [
          { "$ref": "#/components/parameters/SandboxName" },
          { "$ref": "#/components/parameters/Prefer" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResizeRequest" },
              "example": { "size": "50G" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resized and rebooted.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sandbox" }
              }
            }
          },
          "202": { "$ref": "#/components/responses/Accepted" },
          "400": { "$ref": "#/components/responses/Invalid" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "507": { "$ref": "#/components/responses/DiskPoolFull" }
        }
      }
    },
    "/v1/sandboxes/{name}/reboot": {
      "post": {
        "operationId": "reboot",
        "tags": ["sandboxes"],
        "summary": "Cold-boot a sandbox",
        "description": "Discards the memory snapshot and boots the guest fresh. Running processes die; the disk is untouched.",
        "parameters": [
          { "$ref": "#/components/parameters/SandboxName" },
          { "$ref": "#/components/parameters/Prefer" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": {
            "description": "Rebooted.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Sandbox" }
              }
            }
          },
          "202": { "$ref": "#/components/responses/Accepted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "spk_v1.<claims>.<hmac>",
        "description": "An edge session token. Mint one with `ssh ctl@example.com session-token | tr -d '\\r\\n'` (the channel ends the line CRLF, and a stray carriage return in the header is rejected), or `POST /v1/account/tokens` from an existing session."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "spark_session",
        "description": "The same token, as the zone-wide cookie a browser gets from `https://login.example.com`. State-changing calls made with the cookie alone must also prove first-party intent — an `Origin` matching this host, or the header `X-Sparkbox-Console: 1` — because the cookie rides every same-site request. Bearer-authenticated calls need neither."
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "No valid session. Send `Authorization: Bearer <token>`.",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
...
    },
...
  }
}